I took Ratha's files from his program and manually synced them up with the main one. Please DO NOT make seperate programs when adding stuff. Just use the main program because then it makes everything huge pain. TL;DR: Added Ratha's files and DO NOT MAKE SEPERATE PROGRAMS
