+++ Distr Computing Final Project/nbproject/private/private.properties
@@ -1,0 +1,7 @@
+deploy.ant.properties.file=C:\\Users\\OWNER\\AppData\\Roaming\\NetBeans\\8.2\\config\\GlassFishEE6\\Properties\\gfv3-1448918244.properties
+j2ee.platform.is.jsr109=true
+j2ee.server.domain=C:/Users/OWNER/GlassFish_Server/glassfish/domains/domain1
+j2ee.server.home=C:/Users/OWNER/GlassFish_Server/glassfish
+j2ee.server.instance=[C:\\Users\\OWNER\\GlassFish_Server\\glassfish;C:\\Users\\OWNER\\GlassFish_Server\\glassfish\\domains\\domain1]deployer:gfv3ee6wc:localhost:4848
+j2ee.server.middleware=C:/Users/OWNER/GlassFish_Server
+user.properties.file=C:\\Users\\OWNER\\AppData\\Roaming\\NetBeans\\8.2\\build.properties
