--- distr-final-proj-master/Distr Computing Final Project/nbproject/genfiles.properties
@@ -1,8 +1,0 @@
-build.xml.data.CRC32=9ef7a266
-build.xml.script.CRC32=40626705
-build.xml.stylesheet.CRC32=651128d4@1.68.1.1
-# This file is used by a NetBeans-based IDE to track changes in generated files such as build-impl.xml.
-# Do not edit this file. You may delete it but then the IDE will never regenerate such files for you.
-nbproject/build-impl.xml.data.CRC32=9ef7a266
-nbproject/build-impl.xml.script.CRC32=47e8f7bf
-nbproject/build-impl.xml.stylesheet.CRC32=99ea4b56@1.68.1.1
