--- distr-final-proj-master/Distr Computing Final Project/nbproject/private/private.properties
@@ -1,7 +1,0 @@
-deploy.ant.properties.file=C:\\Users\\RATHA\\AppData\\Roaming\\NetBeans\\8.0.2\\config\\GlassFishEE6\\Properties\\gfv3-21537515.properties
-j2ee.platform.is.jsr109=true
-j2ee.server.domain=C:/Users/RATHA/Desktop/glassfish/glassfish/domains/domain1
-j2ee.server.home=C:/Users/RATHA/Desktop/glassfish/glassfish
-j2ee.server.instance=[C:\\Users\\RATHA\\Desktop\\glassfish\\glassfish;C:\\Users\\RATHA\\Desktop\\glassfish\\glassfish\\domains\\domain1]deployer:gfv3ee6wc:localhost:4848
-j2ee.server.middleware=C:/Users/RATHA/Desktop/glassfish
-user.properties.file=C:\\Users\\RATHA\\AppData\\Roaming\\NetBeans\\8.0.2\\build.properties
