--- distr-final-proj-master/Distr Computing Final Project/nbproject/private/private.xml
@@ -1,13 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project-private xmlns="http://www.netbeans.org/ns/project-private/1">
-    <editor-bookmarks xmlns="http://www.netbeans.org/ns/editor-bookmarks/2" lastBookmarkId="0"/>
-    <open-files xmlns="http://www.netbeans.org/ns/projectui-open-files/2">
-        <group>
-            <file>file:/C:/Users/RATHA/Desktop/CPAN%20252/distr-final-proj-master/Distr%20Computing%20Final%20Project/web/marketLogin.jsp</file>
-            <file>file:/C:/Users/RATHA/Desktop/CPAN%20252/distr-final-proj-master/Distr%20Computing%20Final%20Project/src/java/servlet/printerservlet.java</file>
-            <file>file:/C:/Users/RATHA/Desktop/CPAN%20252/distr-final-proj-master/Distr%20Computing%20Final%20Project/src/java/dao/MarketDao.java</file>
-            <file>file:/C:/Users/RATHA/Desktop/CPAN%20252/distr-final-proj-master/Distr%20Computing%20Final%20Project/web/editmarket.jsp</file>
-            <file>file:/C:/Users/RATHA/Desktop/CPAN%20252/distr-final-proj-master/Distr%20Computing%20Final%20Project/src/java/Services/MarketService.java</file>
-        </group>
-    </open-files>
-</project-private>
