--- distr-final-proj-master/Distr Computing Final Project/src/java/model/Market.java
@@ -1,104 +1,0 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-package model;
-
-/**
- *
- * @author RATHA
- */
-public class Market {
-    
-    int id;
-    String firstName;
-    String lastName;
-    String phoneNo;
-    String email;
-    String userName;
-    String password;
-
-    public Market(int id) {
-        this.id = id;
-    }
-    
-
-    public Market() {
-    }
-
-    public Market(String userName, String password) {
-        this.userName = userName;
-        this.password = password;
-    }
-
-    public Market(int id, String firstName, String lastName, String phoneNo, String email, String userName, String password) {
-        this.id = id;
-        this.firstName = firstName;
-        this.lastName = lastName;
-        this.phoneNo = phoneNo;
-        this.email = email;
-        this.userName = userName;
-        this.password = password;
-    }
-
-    public int getId() {
-        return id;
-    }
-
-    public void setId(int id) {
-        this.id = id;
-    }
-
-    public String getFirstName() {
-        return firstName;
-    }
-
-    public void setFirstName(String firstName) {
-        this.firstName = firstName;
-    }
-
-    public String getLastName() {
-        return lastName;
-    }
-
-    public void setLastName(String lastName) {
-        this.lastName = lastName;
-    }
-
-    public String getPhoneNo() {
-        return phoneNo;
-    }
-
-    public void setPhoneNo(String phoneNo) {
-        this.phoneNo = phoneNo;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-
-    public String getUserName() {
-        return userName;
-    }
-
-    public void setUserName(String userName) {
-        this.userName = userName;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-
-    public void setPassword(String password) {
-        this.password = password;
-    }
-    
-    
-    
-    
-}
