--- distr-final-proj-master/Distr Computing Final Project/web/WEB-INF/web.xml
@@ -1,30 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<web-app version="3.1" xmlns="http://xmlns.jcp.org/xml/ns/javaee" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://xmlns.jcp.org/xml/ns/javaee http://xmlns.jcp.org/xml/ns/javaee/web-app_3_1.xsd">
-    <servlet>
-        <servlet-name>printerservlet</servlet-name>
-        <servlet-class>servlet.printerservlet</servlet-class>
-    </servlet>
-    <servlet-mapping>
-        <servlet-name>printerservlet</servlet-name>
-        <url-pattern>/</url-pattern>
-    </servlet-mapping>
-    <context-param>
-        <param-name>jdbcURL</param-name>
-        <param-value>jdbc:mysql://localhost:3306/printmarketing</param-value>
-    </context-param>
-    
-    <context-param>
-        <param-name>jdbcUserName</param-name>
-        <param-value>root</param-value>
-    </context-param>
-    
-    <context-param>
-        <param-name>jdbcPassword</param-name>
-        <param-value></param-value>
-    </context-param>
-    <session-config>
-        <session-timeout>
-            30
-        </session-timeout>
-    </session-config>
-</web-app>
